+++ localpaas_app/entity/setting.go
@@ -25,7 +25,7 @@ type Setting struct {
 	Status    base.SettingStatus `bun:",nullzero"`
 	Name      string             `bun:",nullzero"`
 	Data      string             `bun:",nullzero"`
-	UpdateVer int64
+	UpdateVer int
 
 	CreatedAt time.Time `bun:",default:current_timestamp"`
 	UpdatedAt time.Time `bun:",default:current_timestamp"`
